refactor: move all_gather
