+++ python/paddle/distributed/__init__.py
@@ -27,8 +27,6 @@ from paddle.distributed.fleet.dataset import InMemoryDataset  # noqa: F401
 from paddle.distributed.fleet.dataset import QueueDataset  # noqa: F401
 from paddle.distributed.fleet.base.topology import ParallelMode  # noqa: F401
 
-from .collective import all_gather  # noqa: F401
-from .collective import all_gather_object  # noqa: F401
 from .collective import barrier  # noqa: F401
 from .collective import split  # noqa: F401
 from .collective import new_group  # noqa: F401
@@ -37,6 +35,8 @@ from .collective import wait  # noqa: F401
 from .communication import (
     stream,
     ReduceOp,
+    all_gather,
+    all_gather_object,
     all_reduce,
     alltoall,
     alltoall_single,
@@ -112,4 +112,5 @@ __all__ = [  # noqa
     "irecv",
     "reduce_scatter",
     "rpc",
+    "stream",
 ]
