+++ python/paddle/distributed/communication/__init__.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from .all_gather import all_gather, all_gather_object
 from .all_reduce import all_reduce
 from .broadcast import broadcast
 from .reduce import reduce, ReduceOp
@@ -24,6 +25,8 @@ from .group import is_initialized, destroy_process_group, get_group
 
 __all__ = [
     "ReduceOp",
+    "all_gather",
+    "all_gather_object",
     "all_reduce",
     "alltoall",
     "alltoall_single",
