+++ python/paddle/fluid/tests/unittests/collective/test_collective_allgather_api.py
@@ -34,8 +34,6 @@ class TestCollectiveAllgatherAPI(TestDistBase):
             "int8",
             "uint8",
             "bool",
-            "complex64",
-            "complex128",
         ]
         for dtype in dtypes_to_test:
             self.check_with_place(
@@ -52,8 +50,6 @@ class TestCollectiveAllgatherAPI(TestDistBase):
             "int8",
             "uint8",
             "bool",
-            "complex64",
-            "complex128",
         ]
         for dtype in dtypes_to_test:
             self.check_with_place(
@@ -64,7 +60,7 @@ class TestCollectiveAllgatherAPI(TestDistBase):
                 dtype=dtype,
             )
 
-    def test_allgatther_nccl_dygraph(self):
+    def test_allgather_nccl_dygraph(self):
         dtypes_to_test = [
             "float16",
             "float32",
@@ -74,8 +70,6 @@ class TestCollectiveAllgatherAPI(TestDistBase):
             "int8",
             "uint8",
             "bool",
-            "complex64",
-            "complex128",
         ]
         if self._nccl_version >= 2100:
             dtypes_to_test.append("bfloat16")
@@ -99,8 +93,6 @@ class TestCollectiveAllgatherAPI(TestDistBase):
             "uint8",
             "bool",
             "bfloat16",
-            "complex64",
-            "complex128",
         ]
         for dtype in dtypes_to_test:
             self.check_with_place(
